rpminfo: Directory identification using mode bits
Currently, if a CPIO archive contains an entry for a directory we
attempt to identify it via nlinks. Since it should have at least two
nlinks if it's a directoy. However, I've stumbled across a case where
this is not true. By using the mode bits we should be able to correctly
identify directories without relying on nlinks.

Fixes: #37

Signed-off-by: John Andersen <johnandersenpdx@gmail.com>