Store one one version of i18n strings
RPM files with multiple translated entries for strings would cause a crash when
parsing, so packages with things like:

Summary: One package
Summary(pt_BR): Um pacote

would invoke the extract_string function with a count > 1.

Fully handling all of the i18n capabilities of RPM requires some re-work on the
header handling, but for the sake of simplicity and to be able to parse those
rpm files, all extra translated strings are discarded and only the first one is
kept.